fix key related unit test.
