+++ cdc/sink/dmlsink/txn/event_test.go
@@ -14,6 +14,7 @@
 package txn
 
 import (
+	"sort"
 	"testing"
 
 	timodel "github.com/pingcap/tidb/parser/model"
@@ -52,4 +53,239 @@ func TestGenKeyListCaseInSensitive(t *testing.T) {
 	second := genKeyList(columns, colInfo, 0, []int{0}, 1)
 
 	require.Equal(t, first, second)
+}
+
+func TestGenKeys(t *testing.T) {
+	t.Parallel()
+	testCases := []struct {
+		txn      *model.SingleTableTxn
+		expected []uint64
+	}{{
+		txn:      &model.SingleTableTxn{},
+		expected: nil,
+	}, {
+		txn: &model.SingleTableTxn{
+			Rows: []*model.RowChangedEvent{
+				{
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
+							Value: 12,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
+							Value: 1,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1, 2}},
+				}, {
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
+							Value: 1,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
+							Value: 21,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1, 2}},
+				},
+			},
+		},
+		expected: []uint64{2072713494, 3710968706},
+	}, {
+		txn: &model.SingleTableTxn{
+			Rows: []*model.RowChangedEvent{
+				{
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 12,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 1,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1}, {2}},
+				}, {
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 1,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 21,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1}, {2}},
+				},
+			},
+		},
+		expected: []uint64{318190470, 2109733718, 2658640457, 2989258527},
+	}, {
+		txn: &model.SingleTableTxn{
+			Rows: []*model.RowChangedEvent{
+				{
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.NullableFlag,
+							Value: nil,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.NullableFlag,
+							Value: nil,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1}, {2}},
+				}, {
+					StartTs:  418658114257813514,
+					CommitTs: 418658114257813515,
+					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
+					PreColumns: []*model.Column{
+						nil,
+						{
+							Name:  "a1",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 1,
+						},
+						{
+							Name:  "a3",
+							Type:  mysql.TypeLong,
+							Flag:  model.BinaryFlag | model.HandleKeyFlag,
+							Value: 21,
+						},
+					},
+					TableInfo: &model.TableInfo{
+						TableInfo: &timodel.TableInfo{
+							Columns: []*timodel.ColumnInfo{
+								nil,
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+								{
+									FieldType: *types.NewFieldTypeWithCollation(mysql.TypeLong, "utf8mb4_bin", 10),
+								},
+							},
+						},
+					},
+					IndexColumns: [][]int{{1}, {2}},
+				},
+			},
+		},
+		expected: []uint64{318190470, 2095136920, 2658640457},
+	}}
+	for _, tc := range testCases {
+		keys := genTxnKeys(tc.txn)
+		sort.Slice(keys, func(i, j int) bool { return keys[i] < keys[j] })
+		require.Equal(t, tc.expected, keys)
+	}
 }
