+++ cdc/sink/dmlsink/txn/txn_dml_sink_test.go
@@ -15,12 +15,10 @@ package txn
 
 import (
 	"context"
-	"sort"
 	"sync/atomic"
 	"testing"
 	"time"
 
-	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/cdc/sink/dmlsink"
 	"github.com/pingcap/tiflow/cdc/sink/tablesink/state"
@@ -104,137 +102,4 @@ func TestTxnSinkNolocking(t *testing.T) {
 	time.Sleep(time.Second)
 	require.Equal(t, uint32(100), atomic.LoadUint32(&handled))
 	sink.Close()
-}
-
-func TestGenKeys(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		txn      *model.SingleTableTxn
-		expected []uint64
-	}{{
-		txn:      &model.SingleTableTxn{},
-		expected: nil,
-	}, {
-		txn: &model.SingleTableTxn{
-			Rows: []*model.RowChangedEvent{
-				{
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
-						Value: 12,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
-						Value: 1,
-					}},
-					IndexColumns: [][]int{{1, 2}},
-				}, {
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
-						Value: 1,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.MultipleKeyFlag | model.HandleKeyFlag,
-						Value: 21,
-					}},
-					IndexColumns: [][]int{{1, 2}},
-				},
-			},
-		},
-		expected: []uint64{2072713494, 3710968706},
-	}, {
-		txn: &model.SingleTableTxn{
-			Rows: []*model.RowChangedEvent{
-				{
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 12,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 1,
-					}},
-					IndexColumns: [][]int{{1}, {2}},
-				}, {
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 1,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 21,
-					}},
-					IndexColumns: [][]int{{1}, {2}},
-				},
-			},
-		},
-		expected: []uint64{318190470, 2109733718, 2658640457, 2989258527},
-	}, {
-		txn: &model.SingleTableTxn{
-			Rows: []*model.RowChangedEvent{
-				{
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.NullableFlag,
-						Value: nil,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.NullableFlag,
-						Value: nil,
-					}},
-					IndexColumns: [][]int{{1}, {2}},
-				}, {
-					StartTs:  418658114257813514,
-					CommitTs: 418658114257813515,
-					Table:    &model.TableName{Schema: "common_1", Table: "uk_without_pk", TableID: 47},
-					PreColumns: []*model.Column{nil, {
-						Name:  "a1",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 1,
-					}, {
-						Name:  "a3",
-						Type:  mysql.TypeLong,
-						Flag:  model.BinaryFlag | model.HandleKeyFlag,
-						Value: 21,
-					}},
-					IndexColumns: [][]int{{1}, {2}},
-				},
-			},
-		},
-		expected: []uint64{318190470, 2095136920, 2658640457},
-	}}
-	for _, tc := range testCases {
-		keys := genTxnKeys(tc.txn)
-		sort.Slice(keys, func(i, j int) bool { return keys[i] < keys[j] })
-		require.Equal(t, tc.expected, keys)
-	}
 }
